+++ assets/static/js/school/init.js
@@ -4,12 +4,36 @@
         klp.router.init();
         klp.tabs.init();
         var $infoXHR = klp.api.do("schools/school/" + SCHOOL_ID, {geometry: 'yes'});
+
+        // FIX THIS LATER
+        $('#map-canvas').css('zIndex', 1);
         $infoXHR.done(function(data) {
-            console.log("info", data);
+            var markerLatlng = L.latLng(data.geometry.coordinates[1], data.geometry.coordinates[0]);
+            var map = L.map('map-canvas', {
+                touchZoom: false,
+                scrollWheelZoom: false,
+                doubleClickZoom: false,
+                boxZoom: false,
+                zoomControl: false,
+                attributionControl: false
+            }).setView(markerLatlng, 15);
+
+            L.tileLayer('http://geo.klp.org.in/osm/{z}/{x}/{y}.png', {
+                attribution: 'OpenStreetMap, OSM-Bright'
+            }).addTo(map);
+
+            console.log('markerLatlng', markerLatlng);
+            var marker = L.geoJson(data, {
+                pointToLayer: function(feature, latLng) {
+                    return L.marker(latLng, {icon: klp.utils.mapIcon(data.properties.type.name)});
+                }
+            }).addTo(map);
+
+            map.panBy([0,50], {animate: true, duration: 0.50});
+
             var tpl = swig.compile($('#tpl-school-info').html());
             //var html = swig.render($('#tpl-school-info').html(), {locals: data});
             var html = tpl(data.properties);
-            console.log(html);
             $('#school-info-wrapper').html(html);
         });
         klp.router.start();