+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 Release 0.3:
  - Updated Django version to 1.7c2 [pip-install]
  - Added django-compressor for js / css minification [pip-install]
+ - Modified SchoolDetails Materialized View [run-sql]
 
 Release 0.2: July 26: 5c28f9cbc1aad0d1653401c671285012fd6fa882
  - User auth working